Get user input
